requirements.txt
pytest==7.1.2
requests==2.27.1
selenium==4.1.3
webdriver-manager==3.5.4
pre-commit==2.18.1
Faker==13.7.0
allure-pytest==2.9.45
